Restructured kustomization.yaml test inputs to properly generate outputs.  Rewrote helpers for identifying kustomize projects. Reverting changes to preprocessing test helpers.  Reorganized methods to better suit maintenance.  Removed unneeded paths.
