+++ pkg/safeguards/preprocessing/preprocessing_helpers.go
@@ -3,73 +3,11 @@ package preprocessing
 import (
 	"fmt"
 	"os"
-	"path/filepath"
-	"strings"
 
-	"github.com/Azure/draft/pkg/safeguards"
-	log "github.com/sirupsen/logrus"
-	"gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v3"
 	"helm.sh/helm/v3/pkg/chart"
-	"helm.sh/helm/v3/pkg/chart/loader"
 	"helm.sh/helm/v3/pkg/chartutil"
-	"helm.sh/helm/v3/pkg/engine"
-	"sigs.k8s.io/kustomize/api/krusty"
-	"sigs.k8s.io/kustomize/api/types"
-	"sigs.k8s.io/kustomize/kyaml/filesys"
 )
-
-// Given a Helm chart directory or file, renders all templates and writes them to the specified directory
-func RenderHelmChart(isFile bool, mainChartPath, tempDir string) ([]safeguards.ManifestFile, error) {
-	if isFile { // Get the directory that the Chart.yaml lives in
-		mainChartPath = filepath.Dir(mainChartPath)
-	}
-
-	mainChart, err := loader.Load(mainChartPath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to load main chart: %s", err)
-	}
-
-	loadedCharts := make(map[string]*chart.Chart) // map of chart path to chart object
-	loadedCharts[mainChartPath] = mainChart
-
-	// Load subcharts and dependencies
-	for _, dep := range mainChart.Metadata.Dependencies {
-		// Resolve the chart path based on the main chart's directory
-		chartPath := filepath.Join(mainChartPath, dep.Repository[len("file://"):])
-		chartPath = filepath.Clean(chartPath)
-
-		subChart, err := loader.Load(chartPath)
-		if err != nil {
-			return nil, fmt.Errorf("failed to load chart: %s", err)
-		}
-		loadedCharts[chartPath] = subChart
-	}
-
-	var manifestFiles []safeguards.ManifestFile
-	for chartPath, chart := range loadedCharts {
-		valuesPath := filepath.Join(chartPath, "values.yaml") // Enforce that values.yaml must be at same level as Chart.yaml
-		mergedValues, err := getValues(chart, valuesPath)
-		if err != nil {
-			return nil, fmt.Errorf("failed to load values: %s", err)
-		}
-		e := engine.Engine{Strict: true}
-		renderedFiles, err := e.Render(chart, mergedValues)
-		if err != nil {
-			return nil, fmt.Errorf("failed to render chart: %s", err)
-		}
-
-		// Write each rendered file to the output directory with the same name as in templates/
-		for renderedPath, content := range renderedFiles {
-			outputFilePath := filepath.Join(tempDir, filepath.Base(renderedPath))
-			if err := os.WriteFile(outputFilePath, []byte(content), 0644); err != nil {
-				return nil, fmt.Errorf("failed to write manifest file: %s", err)
-			}
-			manifestFiles = append(manifestFiles, safeguards.ManifestFile{Name: filepath.Base(renderedPath), Path: outputFilePath})
-		}
-	}
-
-	return manifestFiles, nil
-}
 
 // Returns values from values.yaml and release options specified in values.yaml
 func getValues(chart *chart.Chart, valuesPath string) (chartutil.Values, error) {
@@ -113,63 +51,4 @@ func getReleaseOptions(chart *chart.Chart, vals map[string]interface{}) (chartut
 	}
 
 	return mergedValues, nil
-}
-
-func CreateTempDir(p string) string {
-	dir, err := os.MkdirTemp(p, "prefix")
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	return dir
-}
-
-func IsKustomize(p string) bool {
-	return strings.Contains(p, "kustomization.yaml")
-}
-
-func RenderKustomizeManifest(dir, tempDir string) ([]safeguards.ManifestFile, error) {
-	log.Debugf("Rendering kustomization.yaml...")
-
-	kustomizeFS := filesys.MakeFsOnDisk()
-
-	// Create a new Kustomize build options
-	options := &krusty.Options{
-		Reorder:           "none",
-		AddManagedbyLabel: true,
-		LoadRestrictions:  types.LoadRestrictionsRootOnly,
-		PluginConfig:      &types.PluginConfig{},
-	}
-
-	// Create a new Kustomize build object
-	k := krusty.MakeKustomizer(options)
-
-	// Run the build to generate the manifests
-	resMap, err := k.Run(kustomizeFS, dir)
-	if err != nil {
-		return nil, fmt.Errorf("Error building manifests: %s\n", err.Error())
-	}
-
-	// Output the manifests
-	var manifestFiles []safeguards.ManifestFile
-	for _, res := range resMap.Resources() {
-		yamlRes, err := res.AsYAML()
-		if err != nil {
-			return nil, fmt.Errorf("Error converting resource to YAML: %s\n", err.Error())
-		}
-
-		outputRenderPath := filepath.Join(tempDir, res.GetName()) + ".yaml"
-		err = kustomizeFS.WriteFile(outputRenderPath, yamlRes)
-		if err != nil {
-			return nil, fmt.Errorf("Error writing yaml resource: %s\n", err.Error())
-		}
-
-		// write yamlRes to dir
-		manifestFiles = append(manifestFiles, safeguards.ManifestFile{
-			Name: res.GetName(),
-			Path: outputRenderPath,
-		})
-	}
-
-	return manifestFiles, nil
 }
