--- pkg/safeguards/preprocessing/testdata/helm/resources.yaml
@@ -1,40 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: test-release-deployment
-  namespace: test-namespace
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: my-web-app
-  template:
-    metadata:
-      labels:
-        app: my-web-app
-    spec:
-      containers:
-        - name: nginx
-          image: nginx:stable
----
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: test-release-configmap
-  namespace: test-namespace
-data:
-  config.yaml: |
-    replicaCount: 1
-    image:
-      repository: nginx
-      tag: stable
-      pullPolicy: IfNotPresent
-    service:
-      type: ClusterIP
-      port: 80
-    ingress:
-      enabled: true
-      hostname: example.com
-      annotations:
-      tls: false
-      tlsSecret:
