--- pkg/safeguards/preprocessing/testdata/helm/service-2.yaml
@@ -1,11 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: test-release-service2
-  namespace: test-namespace
-spec:
-  type: ClusterIP
-  ports:
-    - port: 80
-  selector:
-    app: my-web-app
