+++ pkg/safeguards/tests/testmanifests/kustomize/base/deployment.yaml
@@ -1,24 +1,23 @@
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  name: {{APPNAME}}
+  name: my-app
   labels:
-    app: {{APPNAME}}
-    kubernetes.azure.com/generator: {{GENERATORLABEL}}
-  namespace: {{NAMESPACE}}
+    app: my-app
+  namespace: my-namespace
 spec:
   replicas: 1
   selector:
     matchLabels:
-      app: {{APPNAME}}
+      app: my-app
   template:
     metadata:
       labels:
-        app: {{APPNAME}}
+        app: my-app
     spec:
       containers:
-        - name: {{APPNAME}}
-          image: {{IMAGENAME}}:{{IMAGETAG}}
+        - name: my-app
+          image: nginx:latest
           imagePullPolicy: Always
           ports:
-            - containerPort: {{PORT}}
+            - containerPort: 80
