+++ pkg/safeguards/tests/testmanifests/kustomize/base/service.yaml
@@ -1,15 +1,13 @@
 apiVersion: v1
 kind: Service
 metadata:
-  name: {{APPNAME}}
-  namespace: {{NAMESPACE}}
-  labels:
-    kubernetes.azure.com/generator: {{GENERATORLABEL}}
+  name: my-service
+  namespace: my-namespace
 spec:
   type: LoadBalancer
   selector:
-    app: {{APPNAME}}
+    app: my-app
   ports:
     - protocol: TCP
-      port: {{SERVICEPORT}}
-      targetPort: {{PORT}}
+      port: 4040
+      targetPort: 80
