+++ pkg/safeguards/tests/testmanifests/kustomize/overlays/production/kustomization.yaml
@@ -1,7 +1,10 @@
 namePrefix: production-
-namespace: {{NAMESPACE}}
+namespace: my-namespace
 resources:
-  - ../../base
-patchesStrategicMerge:
-  - deployment.yaml
-  - service.yaml
+- ../../base
+apiVersion: kustomize.config.k8s.io/v1beta1
+kind: Kustomization
+patches:
+- path: deployment.yaml
+- path: service.yaml
+- path: namespace.yaml
