+++ pkg/safeguards/tests/testmanifests/kustomize/overlays/production/namespace.yaml
