+++ pkg/safeguards/tests/testmanifests/kustomize/overlays/production/service.yaml
@@ -1,9 +1,7 @@
 apiVersion: v1
 kind: Service
 metadata:
-  name: {{APPNAME}}
-  namespace: {{NAMESPACE}}
-  labels:
-    kubernetes.azure.com/generator: {{GENERATORLABEL}}
+  name: my-service
+  namespace: my-namespace
 spec:
   type: LoadBalancer
